Adapters: LoadAdaptersJob will load all adapters
Once Manager::loadAdapters is called, all adapters will be loaded
and Manager::adapterAdded signal will start emitting (only for newly
added adapters after loadAdapters call).
It is needed to have all adapters loaded in order to be able to emit
usableAdapterChanged signal accordingly.
If the adapters are not loaded, the usableAdapterChanged would have
been emitted first with null adapter and then emitted again with
the found usable adapter (async -> with a little delay), which is
probably not what we want.