Jobs: Correctly emit result signal when deleting unfinished job
It would be possible to never emit result signal if the job was deleted
while running. It can happen when deleting running job, or it may happen
when device being loaded gets removed.

Jobs now have new 2 properties running and finished and there is also
a method to kill the job.

Added autotest for the case of deleting running jobs