Force QDBus to cache owner of org.bluez to prevent unnecessary blocking calls
Connecting to non-existant signal to dummy slot before constructing
any QDBusAbstractInterface's will force QDBus to cache owner of
org.bluez (this is one blocking call) and thus preventing to do
any more blocking calls when constructing QDBusAbstractInterface's.

This is really hacky, but saves ~5 blocking calls on startup.
There is now only one blocking call (to org.freedesktop.DBus) at
startup, which is still not ideal, but that seems to be the best
we can do.