Don't ignore interface name in DBus properties changed signal
Fixes issue when disconnecting MediaPlayer profile and Connected
property on org.bluez.MediaControl1 interface changes, then it is
interpreted as a device is disconnected, which is not true.

This is only a minor issue, it requires to disconnect a single profile
on device (and the profile being Services::AudioVideoRemoteControl)
which is not done anywhere in Bluedevil.