Fix determining global state of rfkill switch
Instead of using the least restrictive state of a Bluetooth
device (if you have more devices and one of them is not blocked,
then it would be reported as not blocked) use the most restrictive.

On some laptops there are a virtual Bluetooth devices in rfkill
and to get Bluetooth device up you need to unblock all devices.