isBluetoothOperational now also depends on unblocked rfkill
Even though it is possible that Bluetooth is reported as blocked
and there is an operational adapter, there would also be another
adapter that is powered off and blocked through rfkill, thus unable
to be powered on without unblocking.
It is safer to just require unblocked rfkill for all Bluetooth devices.