Fix property changes being missed immediately after an obejct is added
Fix race condition when property changes may be missed if the property
is changed immediately after the object is created.
The issue was that the connection to PropertyChanged signal was
created only after interfacesAdded signal was fired, which may have
already been too late.
This fixes it with connecting to PropertyChanged signal on all paths
in Manager::init().

BUG: 377405

Differential Revision: https://phabricator.kde.org/D5550