Add functions that allow setting the discovery filter based on Bluez Adapter APIs (see https://git.kernel.org/pub/scm/bluetooth/bluez.git/tree/doc/adapter-api.txt).
There are two new functions for the Adapter class - setDiscoveryFilter(...) and getDiscoveryFilters().
The former allows setting discovery filter settings as a QVariantMap of options. The latter queries
the Bluez adapter object for the array of available filters that are supported by the adapter and returns
an array of strings that are permitted to be keys in QVariantMap passed to setDiscoveryFilter(...).

PendingCall is extended to handle array{string} D-Bus type (QStringList) to allow returning it from
getDiscoveryFilters() function.